Add ETags for immutable directories
Like immutable files, the ETag is based on the storage index.
However, since a directory is a special interpretation of a file, it is
distinguished from the file by prepending "DIR-IMM:" onto the start of
the ETag, and adding -representation on the end (where -representation
is the ?t= argument, json, info, etc).

It also checks the return of setETag and avoids generating a
representation if the client already has it.